+++ OpenOversight/app/db_repository/manage.py
