+++ OpenOversight/app/models.py
@@ -1,4 +1,5 @@
-from . import db
+from flask_sqlalchemy import SQLAlchemy
+db = SQLAlchemy()
 
 class Officer(db.Model):
     __tablename__ = 'officers'
