+++ OpenOversight/migrate_db.py
@@ -1,7 +1,7 @@
 import imp
 from migrate.versioning import api
 
-from app import db
+from app.models import db
 from config import SQLALCHEMY_DATABASE_URI
 from config import SQLALCHEMY_MIGRATE_REPO
