+++ create_db.py
@@ -2,16 +2,17 @@
 
 from migrate.versioning import api
 from migrate.exceptions import DatabaseAlreadyControlledError
-from OpenOversight.app import create_app, db, models
+from OpenOversight.app import create_app, models
+from OpenOversight.app.models import db
 app = create_app('development')
 db.app = app
 import os.path
 
 try:
-    db.create_all()
 
     SQLALCHEMY_MIGRATE_REPO = app.config['SQLALCHEMY_MIGRATE_REPO']
     SQLALCHEMY_DATABASE_URI = app.config['SQLALCHEMY_DATABASE_URI']
+    db.create_all()
 
     if not os.path.exists(SQLALCHEMY_MIGRATE_REPO):
         api.create(SQLALCHEMY_MIGRATE_REPO, 'database repository')
